% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importrefldata.R
\name{importrefldata}
\alias{importrefldata}
\title{Import Reflectance Data}
\usage{
importrefldata(database, password, tablename, lake)
}
\arguments{
\item{database}{Name of Database}

\item{password}{Password for 'root' user}

\item{lake}{Body of Water of Interest}
}
\value{
Dataframe of reflectance values
}
\description{
Reads in Reflectance Data from MySQL Database. It is important that the database have columns for RowID, Blue, Green, Red, NIR, SWIR1, SWIR2,
StationID, ImageDate, Lake, Sensor, and CloudMask. (It may have more columns).
}

