% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeldev.R
\name{modeldev}
\alias{modeldev}
\title{Model Development}
\usage{
modeldev(data, season)
}
\arguments{
\item{data}{Dataframe of reflectance data and field data}

\item{season}{Vector of months included in seasonal model}
}
\value{
Summary of stepwise regression results and plots of parameters that produce the best R2 and BIC
}
\description{
Performs a stepwise Regression and regression subset for a given calibration dataset
}

